fixing cache update issue
