+++ src/main/java/com/example/anirudh/Service/impl/EmployeeServiceImpl.java
@@ -4,7 +4,7 @@ import com.example.anirudh.Accessor.dao.EmployeeDAO;
 import com.example.anirudh.Exceptions.EmployeeNotFoundException;
 import com.example.anirudh.Service.EmployeeService;
 import com.example.anirudh.Validator.EmployeeServiceValidator;
-import com.example.anirudh.cacheManager.CacheManager;
+import com.example.anirudh.cacheManager.EmployeeCache;
 import com.example.anirudh.model.Employee;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
@@ -24,7 +24,7 @@ public class EmployeeServiceImpl implements EmployeeService {
     private final EmployeeDAO employeeDAO;
     private final EmployeeServiceValidator validate;
     private final ObjectMapper jsonObjectMapper;
-    private final CacheManager employeeCache;
+    private final EmployeeCache employeeCache;
 
     @Override
     public List<Employee> getAllEmployees() {
