+++ src/main/java/com/example/anirudh/cacheManager/CacheManager.java
@@ -1,32 +1,21 @@
 package com.example.anirudh.cacheManager;
 
-import com.example.anirudh.Accessor.dao.EmployeeDAO;
-import com.example.anirudh.model.Employee;
-import com.google.common.cache.LoadingCache;
-import lombok.NonNull;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 
-import java.util.HashMap;
-import java.util.List;
-import java.util.concurrent.ScheduledThreadPoolExecutor;
-import java.util.function.Function;
-import java.util.stream.Collectors;
-
 @Slf4j
-public class CacheManager implements InMemoryCache {
+public class CacheManager {
 
-    @NonNull
-    private final LoadingCache<Integer, Employee> employeeCache;
+    private final EmployeeCache employeeCache;
 
-    @NonNull
-    private final EmployeeDAO employeeDAO;
+    private final CacheUpdateTask cacheUpdateTask;
 
     @Autowired
-    public CacheManager(@NonNull EmployeeDAO employeeDAO, @NonNull LoadingCache<Integer, Employee> employeeCache) {
+    public CacheManager(EmployeeCache employeeCache, CacheUpdateTask cacheUpdateTask) {
         this.employeeCache = employeeCache;
-        this.employeeDAO = employeeDAO;
+        this.cacheUpdateTask = cacheUpdateTask;
         buildCache();
+        cacheUpdateTask.setupCacheUpdate();
     }
 
     private void buildCache() {
@@ -38,35 +27,7 @@ public class CacheManager implements InMemoryCache {
     private void initCache() {
         log.info("building employee Cache");
         final long startTime = System.currentTimeMillis();
-        setAll();
+        employeeCache.setAll();
         log.info("employee cache building completed in {} ms", System.currentTimeMillis() - startTime);
-    }
-
-    @Override
-    public void setAll() {
-        List<Employee> employeeList = employeeDAO.findAll();
-        employeeCache.putAll(employeeList.stream()
-                .collect(Collectors.toMap(Employee::getId, Function.identity())));
-    }
-
-    @Override
-    public void set(Employee employee) {
-        employeeCache.put(employee.getId(), employee);
-    }
-
-    @Override
-    public List<Employee> getAll() {
-        HashMap employeeMap = new HashMap(employeeCache.asMap());
-        return employeeMap.values().stream().toList();
-    }
-
-    @Override
-    public Employee get(Employee employee) {
-        return employeeCache.getIfPresent(employee.getId());
-    }
-
-    @Override
-    public void update() {
-        setAll();
     }
 }
