+++ src/main/java/com/example/anirudh/cacheManager/CacheUpdateTask.java
@@ -9,12 +9,11 @@ import java.util.concurrent.TimeUnit;
 @RequiredArgsConstructor(onConstructor = @__(@Autowired))
 public class CacheUpdateTask {
 
-    private final CacheManager cacheManager;
-
+    private final EmployeeCache employeeCache;
     private final ScheduledThreadPoolExecutor threadPoolExecutor = new ScheduledThreadPoolExecutor(5);
 
     public void setupCacheUpdate() {
         threadPoolExecutor.scheduleAtFixedRate(
-                new CacheUpdateRunnable(cacheManager), 1800, 300, TimeUnit.SECONDS);
+                new CacheUpdateRunnable(employeeCache), 300, 300, TimeUnit.SECONDS);
     }
 }
